+++ Application.cpp
@@ -101,4 +101,4 @@ namespace Coda {
 		Coda::Runtime::Interpreter inter = Coda::Runtime::Interpreter();
 		return inter.evaluateProgram(program, env);
 	}
-}
+}