+++ Frontend/Tokens/TokenType.h
@@ -7,6 +7,7 @@ namespace Coda {
 		*/
 		enum class TokenType {
 			INVALID,
+			
 			// Literal Types
 			BOOL,
 			BYTE,
@@ -38,17 +39,6 @@ namespace Coda {
 			CLOSE_BRACKET,
 			OPEN_BRACE,
 			CLOSE_BRACE,
-
-			// Logical Operators
-			EQUALS,
-			NOT_EQUALS,
-			NOT,
-			AND,
-			OR,
-			LESS_THAN,
-			GREATER_THAN,
-			LESS_THAN_OR_EQUAL_TO,
-			GREATER_THAN_OR_EQUAL_TO,
 
 			// End of File
 			END_OF_FILE,