+++ Runtime/RuntimeValue/Value.h
@@ -35,8 +35,13 @@ namespace Coda {
 				: type(other->type), value(other->value), startPosition(other->startPosition), endPosition(other->endPosition), properties(other->properties) {}
 			Value(Type type, Error::Position start, Error::Position end)
 				: type(type), value(), startPosition(start), endPosition(end) {}
-			Value(Type type, std::string value="", Error::Position start = Error::Position(), Error::Position end = Error::Position())
+			Value(Type type, std::string value = "", Error::Position start = Error::Position(), Error::Position end = Error::Position())
 				: type(type), value(value), startPosition(start), endPosition(end) {}
+
+		public:
+			static bool isTruthy(const Value& value);
+
+		public:
 			friend std::ostream& operator<<(std::ostream& os, const Value& value);
 		};